% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{add_site_sequence}
\alias{add_site_sequence}
\title{Add a column with amino acid sequence around a PTM}
\usage{
add_site_sequence(
  obj,
  proteome_fasta,
  master_protein_col = "Master.Protein.Accessions"
)
}
\arguments{
\item{obj}{\code{data.frame} with PD output at PSM/peptide level}

\item{proteome_fasta}{\code{character} Filepath for proteome fasta}

\item{master_protein_col}{\code{character} Column name for master protein}
}
\value{
\code{data.frame}
}
\description{
Add a column with amino acid sequence around a PTM. Value will be
NA if peptide maps to multiple proteins or has multiple PTMs. If padding extends
outside the protein AA sequence, padding will be extended with '_'. The PTM-
centered AA sequence is useful to integrate external databases. Input
is PD output at PSM/peptide level having been passed through \code{parse_PTM_scores}
to add 'filtered_pos' column
}
