% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{addPeptidePositions}
\alias{addPeptidePositions}
\title{Add a column describing the position(s) of the peptide sequence with respect
to the protein}
\usage{
addPeptidePositions(
  obj,
  proteome_fasta,
  master_protein_col = "Master.Protein.Accessions"
)
}
\arguments{
\item{obj}{\code{data.frame} with PD output at PSM/peptide level}

\item{proteome_fasta}{\code{character} Filepath for proteome fasta}

\item{master_protein_col}{\code{character} Column name for master protein}
}
\value{
\code{data.frame}
}
\description{
Identify the position(s) of the peptide sequence in the protein.
Where a sequence has multiple master proteins, the peptide position is
undefined (NA).
}
