% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_crap_fasta.R
\name{append_crap_fasta}
\alias{append_crap_fasta}
\title{Append sequences to end of cRAP FASTA}
\usage{
append_crap_fasta(file, crap_file, add_crap = TRUE, add_crap_start = 1)
}
\arguments{
\item{file}{\code{character}, file path of FASTA to append}

\item{crap_file}{\code{character}, file path of existing cRAP FASTA to append to}

\item{add_crap}{add_crap \code{logical}, should cRAP001, cRAP002, etc. be appended to the
sequence headers in the FASTA to append? Default is \code{TRUE}}

\item{add_crap_start}{\code{numeric}, what number should the cRAP00x start at?
Default is 1.}
}
\value{
Returns the existing cRAP FASTA file with some more sequences added
to the end.
}
\description{
This function is used to add sequences from a FASTA file onto
the end of an existing cRAP FASTA file.
}
\examples{
# Add some commerical protease sequences onto the end of CCP cRAP FASTA
\dontrun{
append_crap_fasta(
  file = "commercial-proteases.fasta",
  crap_file = "ccp_crap_2021-01.fasta",
  add_crap = TRUE
)
}

}
