% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_fasta.R
\name{append_fasta}
\alias{append_fasta}
\title{Add sequences to end of a FASTA file}
\usage{
append_fasta(file1, file2, is_crap = FALSE, crap_start = 1)
}
\arguments{
\item{file1}{\code{character}, file path of FASTA to append.}

\item{file2}{\code{character}, file path of FASTA to append to.}

\item{is_crap}{\code{logical}, should cRAP numbers e.g. cRAP001, cRAP002, etc.
be added to sequence headers of file1? Default is \code{FALSE}.}

\item{crap_start}{\code{numeric}, what number should the cRAPxxx start at?
Default is 1 which will produce: cRAP001.}
}
\value{
Overwrites FASTA file2 with some more sequences added
to the end.
}
\description{
This function is used to add sequences from a FASTA file (file1)
onto the end of another FASTA file (file2).

If file2 is a cRAP database, then you can optionally add cRAP numbers to the
headers of file1, starting at the desired number e.g. cRAP127.
}
\examples{
# Add some commercial protease sequences onto the end of CCP cRAP FASTA
\dontrun{
append_fasta(
  file1 = "commercial-proteases.fasta",
  file2 = "2021-01_CCP_cRAP.fasta",
  is_crap = TRUE,
  crap_start = 128
)
}

}
