% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{center_normalise_to_ref}
\alias{center_normalise_to_ref}
\title{Center-median normalise the expression matrix in an MSnSet using medians
from a reference dataset}
\usage{
center_normalise_to_ref(obj, medians)
}
\arguments{
\item{obj}{\code{MSnSet}. Contains PSMs.}

\item{medians}{\verb{vector, numeric}. Sample medians from reference dataset}
}
\value{
Returns an \code{MSnSet} with the expression matrix column center-median
normalised
}
\description{
Center-median normalisation is a simple normalisation method
that is appropriate for relative abundance proteomics such as isobaric tagging.
This can be achieved with \code{MSnbase::normalise(method='diff.center')}. However,
for some experimental designs, the normalisation should be against the medians
in another dataset. For example, for PTM studies, one may wish to isobaric
tag samples, pool, and then PTM-enriched, with the enriched sample quantified
in a separate run to the non-enriched (total) sample. In this case, it may make
more sense to center-median normalise the PTM-enriched samples using the median
from the total samaples.
}
