% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_runs.R
\name{compare_proteins}
\alias{compare_proteins}
\title{Compare two protein accession strings}
\usage{
compare_proteins(proteins1, proteins2, sep1 = "; ", sep2 = "; ")
}
\arguments{
\item{proteins1}{\code{string}, protein acession(s), set 1}

\item{proteins2}{\code{string}, protein accession(s), set 2}

\item{sep1}{\code{string}, delimiter for proteins1. Default is \code{'; '}.}

\item{sep2}{\code{string}, delimiter for proteins2. Default is \code{'; '}.}
}
\value{
Returns \code{TRUE} if the proteins1 and proteins2 contain the same set
of protein accessions.
}
\description{
Check if two protein accession strings are the same,
allowing for:
\enumerate{
\item multiple accessions (default separated by '; ')
\item multiple accessions in any order
}
}
\keyword{internal}
