% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_runs.R
\name{compare_sequences_per_protein}
\alias{compare_sequences_per_protein}
\title{Identify the proportion of exact matches in protein accessions}
\usage{
compare_sequences_per_protein(seq2new_master_prot, master_prot_col)
}
\arguments{
\item{seq2new_master_prot}{\code{data.frame}, output of merging
\code{\link{read_infiles}} & \code{\link{get_sequence2protein}}.}

\item{master_prot_col}{\code{string}, column name for master protein accessions
e.g. \code{"Master.Protein.Accessions"}.}
}
\value{
Prints the proportion of exact matches in the console.
}
\description{
Identify the proportion of exactly matching protein accessions
in a data.frame containing original and updated master protein assignments.
}
\keyword{internal}
