% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_runs.R
\name{compare_single_master}
\alias{compare_single_master}
\title{Identify the proportion of peptides with a single assigned master protein}
\usage{
compare_single_master(seq2new_master_prot, master_prot_col)
}
\arguments{
\item{seq2new_master_prot}{\code{data.frame}, output of merging
\code{\link{read_infiles}} & \code{\link{get_sequence2protein}}.}

\item{master_prot_col}{\code{string}, column name for master protein accessions
e.g. \code{"Master.Protein.Accessions"}.}
}
\value{
Returns a \code{ggplot} showing the number of peptide sequences per
master protein  in the original infiles, and after the consensus master
protein assignment.
}
\description{
Identify the proportion of peptides with a single assigned
master protein in a data.frame containing original and updated master protein
assignments.
}
\keyword{internal}
