% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features_per_protein.R
\name{count_features_per_protein}
\alias{count_features_per_protein}
\title{Count the number of features (rows) per protein}
\usage{
count_features_per_protein(obj, master_prot_col = "Master.Protein.Accessions")
}
\arguments{
\item{obj}{\code{MSnSet}, rows are PSMs or peptides.}

\item{master_prot_col}{\code{string}, column name for master protein accessions.
Default is \code{"Master.Protein.Accessions"}.}
}
\value{
Returns a \code{tibble} where each row is a master protein, with three columns:
\enumerate{
\item \code{sample} = sample name (i.e. column name)
\item \code{master_protein_col} = master protein accession
\item \code{n} = number of features assigned to the master protein
}
}
\description{
Given an MSnSet, determine the number of features (rows) per
protein for each sample (column).

Useful for bias of variance correction when performing differential testing with
\url{https://www.bioconductor.org/packages/release/bioc/html/DEqMS.html}{DEqMS},
as proteins quantified by more PSMs/peptides tend to lower variance.
}
\examples{
# Use a small example TMT dataset included with the camprotR package
df <- psm_tmt_total

# Make an MSnSet
df_exprs <- as.matrix(df[, grep("Abundance", colnames(df))])
colnames(df_exprs) <- gsub("Abundance\\\\.", "", colnames(df_exprs))

df_fData <- df[, grep("Abundance", colnames(df), invert = TRUE)]

psm <- MSnbase::MSnSet(exprs = df_exprs, fData = df_fData)

# Count the number of PSMs per protein
count_features_per_protein(psm, master_prot_col = "Master.Protein.Accessions")

}
