% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_gene2cat.R
\docType{data}
\name{dep_gene2cat}
\alias{dep_gene2cat}
\title{GO terms assigned to proteins in DEP results}
\format{
An object of class \code{list} of length 1728.
}
\usage{
dep_gene2cat
}
\description{
Description.
}
\references{
Insert reference here.
}
\keyword{datasets}
