% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_results.R
\docType{data}
\name{dep_results}
\alias{dep_results}
\title{Results table from DEP package}
\format{
An object of class \code{data.frame} with 1745 rows and 11 columns.
}
\usage{
dep_results
}
\description{
Description.

\describe{
\item{column name}{\code{class}. Description.}
}
}
\references{
Insert reference here.
}
\keyword{datasets}
