% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{determine_offspring_function}
\alias{determine_offspring_function}
\title{Determine GO offspring object}
\usage{
determine_offspring_function(term, ontology)
}
\arguments{
\item{term}{\code{string}. A single GO.ID.}

\item{ontology}{\code{string}. A single ontology, one of: BP, MF, CC.}
}
\value{
Returns an \code{AnnDbBimap} object which maps GO terms (BP, MF, or
CC) to all offspring terms. Each GO term is mapped to a vector of offsrping
GO terms.
}
\description{
For a single GO term, get its ontology and then return the
correct offpsring mapping object for that ontology.
}
\keyword{internal}
