% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ccp_crap.R
\name{download_ccp_crap}
\alias{download_ccp_crap}
\title{Download CCP cRAP FASTA file}
\usage{
download_ccp_crap(file, add_crap = TRUE, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{file}{\code{character}, filepath to save the fasta to e.g. \code{"crap.fasta"}}

\item{add_crap}{\code{logical}, should cRAP001, cRAP002, etc. be appended to the
sequence headers in the FASTA file? Default is \code{TRUE}}

\item{overwrite}{\code{logical}, if the FASTA file already exists should it be
overwritten? Default is \code{FALSE}}

\item{verbose}{\code{logical}, should the function send messages to the console?
Default is \code{TRUE}}
}
\value{
Returns a FASTA saved to disk at the specified file path.
}
\description{
This function is used to download the Cambridge Centre for
Proteomics cRAP FASTA database using sequences from the lastest UniProt release.
}
\examples{
\dontrun{
download_ccp_crap("path/to/file/crap_2021_01.fasta")
}

}
