% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_crap.R
\name{download_crap}
\alias{download_crap}
\title{Download cRAP/contaminant FASTA files}
\usage{
download_crap(
  file,
  type = "ccp",
  add_crap = TRUE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{file}{\code{character}, filepath to save the fasta to e.g. \code{"crap.fasta"}}

\item{type}{\code{character}, type of cRAP database to download e.g. \code{"ccp"}}

\item{add_crap}{\code{logical}, should cRAP001, cRAP002, etc. be appended to the
sequence headers in the FASTA file? Default is \code{TRUE}}

\item{overwrite}{\code{logical}, if the FASTA file already exists should it be
overwritten? Default is \code{FALSE}}

\item{verbose}{\code{logical}, should the function send messages to the console?
Default is \code{TRUE}}
}
\value{
Returns a file saved to disk at the specified filepath
}
\description{
This function is used to download a specific cRAP/contaminants
sequence database to a FASTA file. The different types of cRAP databases
available through this function currently include:
\itemize{
\item \strong{ccp} = the contaminants database used by the Cambridge Centre for
Proteomics. It is largely based off of The Global Proteome Machine's
cRAP database found \href{https://www.thegpm.org/crap/}{here}, with some extra
sequences including commercial proteases added to the end.
}
}
\examples{
\dontrun{
download_crap("path/to/file/crap_2021_01.fasta", type = "ccp")
}

}
