% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{estimate_go_overrep}
\alias{estimate_go_overrep}
\title{Estimate effect size of GO over-representation}
\usage{
estimate_go_overrep(obj, pwf, gene2cat)
}
\arguments{
\item{obj}{\code{data.frame} containing \code{goseq} results as generated by
\code{\link{get_enriched_go}} or \code{\link[goseq]{goseq}}.}

\item{pwf}{\code{data.frame} as used in \code{\link{get_enriched_go}} or
\code{\link[goseq]{goseq}}.}

\item{gene2cat}{\code{data.frame} as used in \code{\link{get_enriched_go}} or
\code{\link[goseq]{goseq}}.}
}
\value{
Returns \code{obj} with an extra column added called \code{adj_overrep}. This
column is calculated for each GO term by:

\code{numDEInCat / numInCat / (avgTermWeight / avgNonTermWeight) / (totalDEFeatures / totalFeatures)}

where:
\itemize{
\item \code{numDEInCat} is the number of differentially expressed genes (aka. proteins)
assigned to that GO term.
\item \code{numInCat} is the total number of genes (aka. proteins) annotated to that
GO term.
\item \code{avgTermWeight} is the average \code{pwf$pwf} value for all the differentially
expressed genes that were assigned to that GO term.
\item \code{avgNonTermWeight} is the average \code{pwf$pwf} for all the other genes supplied
in \code{pwf}.
\item \code{totalDEFeatures} is the total number of differentially expressed genes
indicated in \code{pwf}.
\item \code{totalFeatures} is the total number of genes indicated in \code{pwf}, i.e. the
number of rows.
}
}
\description{
This is a crude function to estimate the effect size of GO
over-representation i.e. we know a term is over-represented, but we want to
estimate the effect size/\emph{how} over-represented it is. This function should
be run after \code{\link{get_enriched_go}}.
}
