% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{expand_terms}
\alias{expand_terms}
\title{Expand data.frame GO terms}
\usage{
expand_terms(go_df, go_col, go2Ancestor)
}
\arguments{
\item{go_df}{\code{data.frame} for a single protein with a column ==
\code{"GO.ID"}.}

\item{go_col}{\code{variable}. Name of column from the data.frame that
contains the GO terms.}

\item{go2Ancestor}{\verb{named list}. Returned by get_all_mappings.
Names == GO.IDs and values == all ancestor GO.IDs.}
}
\value{
Returns a \code{data.frame} containing all ancestor GO terms for a
single protein.
}
\description{
Input is a data.frame with all the GO terms annotated to a single protein in
a single column. Output is a new data.frame with all of the GO terms for
that protein (annotated and ancestor).
}
\keyword{internal}
