% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{get_ancestor_go}
\alias{get_ancestor_go}
\title{Get all ancestor GO terms}
\usage{
get_ancestor_go(
  go_df,
  feature_col = "UNIPROTKB",
  go_col = "GO.ID",
  return_early_debug = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{go_df}{\code{data.frame}. Contains all initial GO terms for proteins
of interest with \verb{columns == [feature_col, go_col]}.}

\item{feature_col}{\code{string}. Name of the column with the features, e.g.
\code{"UNIPROTKB"}.}

\item{go_col}{\code{string}. Name of the column with the GO ids, e.g. \code{"GO.ID"}.}

\item{return_early_debug}{\code{logical}. Stop function early and return a
list of GO terms and ancestor GO terms for debugging purposes.}

\item{verbose}{\code{logical}.}
}
\value{
Returns a \code{data.frame} containing all ancestor GO terms for all
proteins plus GO term descriptions and ontologies.
}
\description{
Given a data.frame with with a column containing GO terms, this function
will output a data.frame with with the ontologies of those GO terms, their
description, and all their ancestor terms.
}
