% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_crap_fasta.R
\name{get_ccp_crap}
\alias{get_ccp_crap}
\title{Get the UniProt sequences in CCP cRAP}
\usage{
get_ccp_crap()
}
\value{
Returns a \verb{character vector} of UniProt accessions.
}
\description{
This function take no inputs and simply outputs a vector of the
UniProt accessions for the sequences in the CCP cRAP FASTA.
}
