% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{get_enriched_go}
\alias{get_enriched_go}
\title{GO term enrichment using goseq}
\usage{
get_enriched_go(
  pwf,
  gene2cat = NULL,
  ...,
  shorten_term = TRUE,
  shorten_lims = c(1L, 30L)
)
}
\arguments{
\item{pwf}{\code{data.frame} with 3 columns (\code{DEgenes} = logical,
\code{bias.data} = numeric/integer, \code{pwf} = numeric) and row names (usually UniProt
accessions, Ensembl gene IDs or similar).
Typically constructed using \code{\link[goseq]{nullp}}.}

\item{gene2cat}{\code{data.frame} with 2 columns containing the mapping between
genes (usually UniProt accessions, Ensembl gene IDs or similar) and GO terms.
Alternatively, a \verb{named list} where the names are genes and each entry is
a \verb{character vector} of GO terms.}

\item{...}{Other arguments to be passed to \code{\link[goseq]{goseq}}.}

\item{shorten_term}{\code{logical}. Should an extra column with a substring of
the output GO terms be added to the output data.frame? Default is \code{TRUE}.}

\item{shorten_lims}{\verb{integer vector} of length 2. The start and stop
coordinates of the substring.}
}
\value{
Returns a \code{data.frame} of over/underrepresented GO terms.
}
\description{
A wrapper function around \code{\link[goseq]{goseq}} to perform
GO term enrichment analysis. See the \code{\link[goseq]{goseq}} documentation
for details. \code{pwf} can be made using \code{\link[goseq]{nullp}}.

Over/underrepresented p-values are automatically
adjusted using \code{method = "BH"}. If \code{gene2cat} is not provided then this
function will default to using the Homo sapiens genome \code{hg19} and will
expect Ensembl gene IDs to have been used to construct the \code{pwf} input.
}
