% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{get_medians}
\alias{get_medians}
\title{Extract the expression matrix column medians from an MSnSet}
\usage{
get_medians(obj)
}
\arguments{
\item{obj}{\code{MSnSet}. Contains PSMs.}
}
\value{
\code{vector} of expression matrix column medians in colnames order
}
\description{
Extract the expression matrix column medians from an MSnSet
}
