% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt_qc_plots.R
\name{get_notch_per_protein}
\alias{get_notch_per_protein}
\title{Summarise the number of sub-notch PSM intensities per protein}
\usage{
get_notch_per_protein(
  obj,
  master_prot_col = "Master.Protein.Accessions",
  notch_upper = 5.75
)
}
\arguments{
\item{obj}{\code{MSnSet} containing PSM-level TMT intensities.}

\item{master_prot_col}{\code{character}. Column name for master proteins.}

\item{notch_upper}{\code{numeric}. Upper boundary of notch.}
}
\value{
\code{data.frame} Summarises # PSMs above/below notch for each protein
}
\description{
The distribution of TMT intensities obtained by Orbitrap has
been observed to have a 'notch' where few values are observed. This function
summarises the number of sub-notch values observed per protein
}
