% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_runs.R
\name{get_parsimony_pep2prot}
\alias{get_parsimony_pep2prot}
\title{Get consensus peptide to master protein mapping across multiple files}
\usage{
get_parsimony_pep2prot(
  infiles,
  seq_col = "Sequence",
  prot_col = "Protein.Accessions",
  master_prot_col = "Master.Protein.Accessions",
  compare_old_new = TRUE
)
}
\arguments{
\item{infiles}{\code{list}, paths for PD peptideGroups.txt files to be read
by \code{\link[utils]{read.delim}}.}

\item{seq_col}{\code{string}, column name for peptide sequence e.g. \code{"Sequence"}.}

\item{prot_col}{\code{string}, column name for protein accessions e.g.
\code{"Protein.Accessions"}.}

\item{master_prot_col}{\code{string}, column name for master protein accessions
e.g. \code{"Master.Protein.Accessions"}.}

\item{compare_old_new}{\code{logical}, compare the consensus master protein
assignments with the original assignments. Default is \code{TRUE}.}
}
\value{
Returns a \code{data.frame} which maps peptide sequences to
consensus master protein accession(s). Each row is a peptide.
}
\description{
Given multiple peptideGroups.txt files, identify the
parsimonious master proteins to account for the all the peptides identified
and (optionally) compare the new and old master protein assignments.

The output of this function can then be used when merging multiple
peptideGroups.txt files (i.e. the outputs of separate PD studies).
}
\examples{
\dontrun{
# The input in this example is 2x peptideGroups.txt outputs from Proteome
# Discoverer. Both files must have the columns (cols) specified. This may require
# re-exporting the peptideGroups.txt file from PD as the "Protein.Accessions"
# column is often not there by default.
get_parsimony_pep2prot(
  list("peptideGroups1.txt", "peptideGroups2.txt"),
  seq_col = "Sequence",
  prot_col = "Protein.Accessions",
  master_prot_col = "Master.Protein.Accessions",
  compare_old_new = TRUE
)

}
}
