% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{get_psm_metrics}
\alias{get_psm_metrics}
\title{Summarise percentage of tag intensities below notch and missing}
\usage{
get_psm_metrics(
  obj,
  notch_lower = 3.75,
  notch_upper = 5.75,
  group_by_sample = FALSE
)
}
\arguments{
\item{obj}{\code{MSnSet}.}

\item{notch_lower}{\code{numeric} Lower boundary of notch}

\item{notch_upper}{\code{numeric} Upper boundary of notch}

\item{group_by_sample}{Group the summaries by sample}
}
\value{
\code{data.frame} object.
}
\description{
Given an \code{MSnSet}, return a \code{data.frame} summarising the
percentage of tag intensities below notch and missing
and the median tag intensity
}
