% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratio.R
\name{get_ratio}
\alias{get_ratio}
\title{Get ratio of two numbers}
\usage{
get_ratio(data, treatment, control, bind = TRUE)
}
\arguments{
\item{data}{\code{data.frame}.}

\item{treatment, control}{\code{variable}. Name of numeric column in the data.frame.}

\item{bind}{\code{logical}. Should the resulting ratios be added to data
as a column? Default is \code{TRUE}.}
}
\value{
Returns \code{data} with 2 new columns: ratio (a numeric column) and
missing (a character column indicating if there was missing values in
\code{treament} and/or \code{control}). Returns a \verb{named list} if \code{bind = FALSE}.
}
\description{
Calculates the ratio of two \code{numeric} columns
in a data.frame, and indicates if there are missing values. Calculation is:
\code{trt - ctr}. Typically used to calculate the ratio of two \code{log2}
transformed SILAC/label-free intensity columns.
}
\examples{
my_data <- data.frame(
  treatment = c(30, NA, 12, NA),
  control = c(20, 12, NA, NA)
)

my_data2 <- get_ratio(my_data, treatment, control)
}
