% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_runs.R
\name{get_sequence2protein}
\alias{get_sequence2protein}
\title{Map peptide sequence to protein across multiple files using
approximate parsimony}
\usage{
get_sequence2protein(seq2prot, seq_col, prot_col)
}
\arguments{
\item{seq2prot}{\code{data.frame}, output from \code{\link{read_infiles}}.}

\item{seq_col}{\code{string}, column name for peptide sequence e.g. \code{"Sequence"}.}

\item{prot_col}{\code{string}, column name for protein accessions e.g.
\code{"Protein.Accessions"}.}

\item{master_prot_col}{\code{string}, column name for master protein accessions
e.g. \code{"Master.Protein.Accessions"}.}
}
\value{
Returns a \code{data.frame} object which uniquely maps peptide sequence
to master protein(s).
}
\description{
Generate a single map of peptide sequences to proteins for
multiple files (e.g. multiple separate peptideGroups.txt PD files) using
an approximate parsimony approach, specifically the least number
of proteins required to account for all observed peptide sequences.
}
\keyword{internal}
