% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{getSequence}
\alias{getSequence}
\title{Get the amino acid sequence around a PTM}
\usage{
getSequence(proteome, protein, ptm_position, pad = 7)
}
\arguments{
\item{proteome}{\code{XStringSet} as generated by \code{Biostrings::readAAStringSet}}

\item{protein}{\code{character} protein ID}

\item{ptm_position}{\code{numeric} position of the PTM in the protein}

\item{pad}{\code{numeric} Number of amino acids around PTM}
}
\value{
\code{character} PSM-centered amino acid sequence
}
\description{
Get the amino acid sequence around a PTM. Will return NA if
peptide maps to multiple proteins or has multiple PTMs. If padding extends
outside the protein AA sequence, padding will be extended with '_'.
}
