% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycoproteins.R
\docType{data}
\name{glycoproteins}
\alias{glycoproteins}
\title{Glycoproteoins}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 163 rows and 2 columns.
}
\usage{
glycoproteins
}
\description{
data.fram with glycoproteins for use in LFQ vignette

\describe{
\item{column name}{\code{class}. Description.}
}
}
\keyword{datasets}
