% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq_oops_rnase.R
\docType{data}
\name{lfq_oops_rnase_psm}
\alias{lfq_oops_rnase_psm}
\title{PSM PD output for LFQ of RNase +/- OOPS experiment}
\format{
An object of class \code{data.frame} with 3673 rows and 42 columns.
}
\usage{
lfq_oops_rnase_psm
}
\description{
OOPS CL vs NC
F17-20 = RNase(-) reps 1-4
F21-24 = RNase(+) reps 1-4

\describe{
\item{column name}{\code{class}. Description.}
}
}
\keyword{datasets}
