% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_crap_fasta.R
\name{make_crap_fasta}
\alias{make_crap_fasta}
\title{Make a cRAP FASTA using UniProt accessions}
\usage{
make_crap_fasta(
  accessions,
  file,
  add_crap = TRUE,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{accessions}{\verb{character vector}, the UniProt accessions to use}

\item{file}{\code{character}, filepath to save the fasta to e.g. \code{"crap.fasta"}}

\item{add_crap}{\code{logical}, should cRAP001, cRAP002, etc. be appended to the
sequence headers in the FASTA file? Default is \code{TRUE}}

\item{overwrite}{\code{logical}, if the FASTA file already exists should it be
overwritten? Default is \code{FALSE}}

\item{verbose}{\code{logical}, should the function send messages to the console?
Default is \code{TRUE}}
}
\value{
Returns a FASTA file saved to disk at the specified file path.
}
\description{
Given a vector of UniProt accessions, this function will:
\enumerate{
\item Download the sequences
\item Print the current UniProt release (you should put this in the FASTA file name)
\item (Optional) add cRAP numbers to the FASTA headers for each sequence
\item Save the sequences in a FASTA file
}
}
\examples{
# specify some UniProt accessions
crap <- get_ccp_crap()

\dontrun{
make_crap_fasta(crap, "crap_2021-01.fasta")
}

}
