% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{match_proteins}
\alias{match_proteins}
\title{Check if two protein IDs are the same allowing for:
\enumerate{
\item multiple IDs (default separated by '; ')
\item mutiple IDs in any order
}}
\usage{
match_proteins(proteins1, proteins2, sep1 = "; ", sep2 = "; ")
}
\arguments{
\item{proteins1}{\code{string} protein(s), set 1}

\item{proteins2}{\code{string} protein(s), set 2}

\item{sep1}{\code{string} delimiter for proteins1}

\item{sep2}{\code{string} delimiter for proteins2}

\item{seq2prot}{\code{data}. Output from \code{\link{read_infiles}}}
}
\value{
Returns an \code{Logical} where \code{TRUE} if the proteins1 and
proteins2 contain the same set of proteins
}
\description{
Check if two protein IDs are the same allowing for:
}
\keyword{internal}
