% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptm.R
\name{parse_PTM_scores}
\alias{parse_PTM_scores}
\title{Parse the PTM probabilities and add new columns with PTM information}
\usage{
parse_PTM_scores(
  obj,
  threshold = 95,
  ptm_col = "PhosphoRS.Best.Site.Probabilities",
  prob_split = "; |: ",
  collapse_delimiter = ";",
  verbose = TRUE
)
}
\arguments{
\item{obj}{\code{data.frame} with PD output at PSM/peptide level}

\item{ptm_col}{\code{character} Columm name for PTM probabilities}

\item{prob_split}{\code{character} regex to split PTM probabilities}

\item{collapse_delimiter}{\code{character} delimiter for multiple values in output columns}

\item{verbose}{Set TRUE to print log of events to console}
}
\value{
\code{data.frame}
}
\description{
Extract PTM information from the PTM probabilities column in
the output from PD and add new columns with this information. Also optionally
prints a summary of how many features (PSMs/peptides) pass a given probability
threshold
}
