% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_oops_3.R
\docType{data}
\name{pep_oops_3}
\alias{pep_oops_3}
\title{PeptideGroups PD output for SILAC labelled U-2 OS cells}
\format{
An object of class \code{data.frame} with 214 rows and 31 columns.
}
\usage{
pep_oops_3
}
\description{
OOPS CL vs NC, H=NC, L=CL.

\describe{
\item{column name}{\code{class}. Description.}
}
}
\references{
Insert reference here when published.
}
\keyword{datasets}
