% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_silac_p0.R
\docType{data}
\name{pep_silac_p0}
\alias{pep_silac_p0}
\title{PeptideGroups PD output for SILAC labelled MOLM-13 cells}
\format{
An object of class \code{data.frame} with 418 rows and 79 columns.
}
\usage{
pep_silac_p0
}
\description{
Passage 0, 100\% light (R0K0).

\describe{
\item{column name}{\code{class}. Description.}
}
}
\references{
Insert reference here when published.
}
\keyword{datasets}
