% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pep_silac_p4.R
\docType{data}
\name{pep_silac_p4}
\alias{pep_silac_p4}
\title{PeptideGroups PD output for SILAC labelled MOLM-13 cells}
\format{
An object of class \code{data.frame} with 470 rows and 79 columns.
}
\usage{
pep_silac_p4
}
\description{
Passage 4, 1:1 light:heavy mixture (R0K0:R10K8).

\describe{
\item{column name}{\code{class}. Description.}
}
}
\references{
Insert reference here when published.
}
\keyword{datasets}
