% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt_qc_plots.R
\name{plot_fraction_below_notch_per_prot}
\alias{plot_fraction_below_notch_per_prot}
\title{Plot the fraction of sub-notch PSM intensities per protein}
\usage{
plot_fraction_below_notch_per_prot(notch_per_protein)
}
\arguments{
\item{notch_per_protein}{\code{data.frame}. Notch summary as generated by
\code{get_notch_per_protein}}
}
\value{
\code{ggplot} stacked bar plot with # sub-notch PSM intensities per protein
}
\description{
The distribution of TMT intensities obtained by Orbitrap has
been observed to have a 'notch' where few values are observed. This function
plots the fraction of sub-notch values observed per protein, as obtained using
\code{get_notch_per_protein}
}
