% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{plot_go}
\alias{plot_go}
\title{Plot selected GO terms}
\usage{
plot_go(
  obj,
  term_col = "term_short",
  ontology_col = "ontology",
  annotate_n = TRUE
)
}
\arguments{
\item{obj}{\code{data.frame} containing \code{goseq} results as generated by
\code{\link{get_enriched_go}} then \code{\link[goseq]{estimate_go_overrep}}.
See below for an example.}

\item{term_col}{\code{character} column name for GO term description.}

\item{ontology_col}{\code{character} column name for GO term ontology.}

\item{annotate_n}{\code{logical} Include the number of features per term
(i.e. numDEInCat column) in the plot? (Default is \code{TRUE})}
}
\value{
Returns a \code{ggplot} object.
}
\description{
This function plots a set of GO terms of interest.
be run after \code{\link{get_enriched_go}} and \code{\link{estimate_go_overrep}}.
To avoid plotting too many terms, you may wish to use
\code{\link{remove_redundant_go}} first too. A warning is shown if you try
and plot more than 100 GO terms.
}
