% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{plot_incorporation}
\alias{plot_incorporation}
\title{Plot annotated histogram of incorporation values}
\usage{
plot_incorporation(
  data,
  incorporation_col = "Incorporation",
  corrected_col = "Incorporation.corrected",
  level = "peptide",
  mix = 0
)
}
\arguments{
\item{data}{\code{data.frame} containing features with column 'incorporation'.
If mix>0, must also contain column 'corrected_incorporation' where incorporation
is calculated using the corrected light peptide intensity}

\item{incorporation_col}{\code{string}. Name of the column containing incorporation values
which can be calculated with \code{\link{get_incorporation}}.}

\item{corrected_col}{\code{string} Optional, but required if mix > 0. Name of the
column containing mix-corrected incorporation values which can be calculated
with \code{\link{get_incorporation}}.}

\item{level}{\code{string}. Name for feature level, either 'peptide' or 'protein'.}

\item{mix}{\code{numeric}. The 'Heavy' to 'Light' mix ratio. Default is 0, otherwise
use a number greater than 0.}
}
\value{
Returns a \code{list} with \code{p} = \code{ggplot} plot and
\code{incorporation_estimates} = \code{list} of incorporation estimates.
}
\description{
SILAC (2-plex) label incorporation is estimated from the set of
observed features (peptides or proteins). If the test sample contains cells
grown in 'Heavy' media only, incorporation is simply the mean observed
incorporation. If the test sample contains a mixture of
'Heavy' and 'Light', the incorporation is best calculated from the median
incorporation using the corrected 'Light' intensity.
}
