% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{plot_label_quant}
\alias{plot_label_quant}
\title{Plot distributions for feature intensities per sample.}
\usage{
plot_label_quant(obj, method = c("box", "density", "histogram"))
}
\arguments{
\item{obj}{\code{MSnSet}.}

\item{method}{\code{string} Plot style. Choice of box, density or histogramplot.}
}
\value{
\code{ggplot} object.
}
\description{
Given an \code{MSnSet}, return a plot
of the feature intensities per sample.
}
