% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt_qc_plots.R
\name{plot_missing_SN}
\alias{plot_missing_SN}
\title{Plot the missing values vs signal:noise}
\usage{
plot_missing_SN(obj, sn_column = "Average.Reporter.SN", bins = 20)
}
\arguments{
\item{obj}{\code{MSnSet} containing PSM level TMT intensities}

\item{sn_column}{\code{character} column name for signal:noise values}

\item{bins}{\code{numeric} Number of bins to plot}
}
\value{
\code{ggplot} stacked bar plot to show S:N vs # missing values
}
\description{
Missing values are more frequent with low signal:noise (S:N).
This function visualises this relationship to aid selection of thresholds for
minimal S:N filtering
}
