% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{plot_quant}
\alias{plot_quant}
\title{Plot distributions for feature intensities per sample.}
\usage{
plot_quant(
  obj,
  method = c("box", "density", "histogram"),
  facet_by_sample = FALSE
)
}
\arguments{
\item{obj}{\code{MSnSet}.}

\item{method}{\code{string} Plot style. Choice of box, density or histogramplot.}

\item{facet_by_sample}{Facet the plot by sample}
}
\value{
\code{ggplot} object.
}
\description{
Given an \code{MSnSet}, return a plot
of the feature intensities per sample.
}
