% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_runs.R
\name{plot_sequences_per_protein}
\alias{plot_sequences_per_protein}
\title{Plot the number of peptides per protein pre/post-consensus master protein
assignment}
\usage{
plot_sequences_per_protein(seq2new_master_prot, seq_col, master_prot_col)
}
\arguments{
\item{seq2new_master_prot}{\code{data.frame}, output of merging output of
\code{\link{read_infiles}} & \code{\link{get_sequence2protein}}.}

\item{seq_col}{\code{string}, column name for peptide sequence e.g. \code{"Sequence"}.}

\item{master_prot_col}{\code{string}, column name for master protein accessions
e.g. \code{"Master.Protein.Accessions"}.}
}
\value{
Returns a \code{ggplot} showing the number of sequences per master protein
in the original infiles, and after the consensus master protein assignment.
}
\description{
Plot the number of peptides per protein pre/post-consensus master protein
assignment
}
\keyword{internal}
