% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psm_data_notch_count}
\alias{psm_data_notch_count}
\title{PSM-level PD output from TMT experiment}
\format{
An MSnSet with fVarLabels:
\describe{
\item{Master.Protein.Accessions}{Protein Accessions}
}
}
\usage{
psm_data_notch_count
}
\description{
PSM-level PD output from TMT experiment
}
\examples{
\dontrun{
 psm_data_notch_count
}
}
\keyword{datasets}
