% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_oops_2.R
\docType{data}
\name{psm_oops_2}
\alias{psm_oops_2}
\title{PSM PD output for SILAC labelled U-2 OS cells}
\format{
An object of class \code{data.frame} with 609 rows and 44 columns.
}
\usage{
psm_oops_2
}
\description{
OOPS CL vs NC, H=CL, L=NC.

\describe{
\item{column name}{\code{class}. Description.}
}
}
\references{
Insert reference here when published.
}
\keyword{datasets}
