% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_oops_3.R
\docType{data}
\name{psm_oops_3}
\alias{psm_oops_3}
\title{PSM PD output for SILAC labelled U-2 OS cells}
\format{
An object of class \code{data.frame} with 2035 rows and 44 columns.
}
\usage{
psm_oops_3
}
\description{
OOPS CL vs NC, H=NC, L=CL.

\describe{
\item{column name}{\code{class}. Description.}
}
}
\references{
Insert reference here when published.
}
\keyword{datasets}
