% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_tmt_phospho.R
\docType{data}
\name{psm_tmt_phospho}
\alias{psm_tmt_phospho}
\title{PSM-level PD output for phospho TMT10-plex data}
\format{
An object of class \code{data.frame} with 5000 rows and 52 columns.
}
\usage{
psm_tmt_phospho
}
\description{
Description.

\describe{
\item{column name}{\code{class}. Description.}
}
}
\references{
Insert reference here when published.
}
\keyword{datasets}
