% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psm_tmt_total.R
\docType{data}
\name{psm_tmt_total}
\alias{psm_tmt_total}
\title{PSM-level PD output for total proteome TMT10-plex data}
\format{
An object of class \code{data.frame} with 5000 rows and 50 columns.
}
\usage{
psm_tmt_total
}
\description{
Description.

\describe{
\item{column name}{\code{class}. Description.}
}
}
\references{
Insert reference here when published.
}
\keyword{datasets}
