% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{psm_to_peptide_style_modifications}
\alias{psm_to_peptide_style_modifications}
\title{Transformes PSM-level Modifications annotation style to peptide-level style}
\usage{
psm_to_peptide_style_modifications(psm_style_modifications)
}
\arguments{
\item{obj}{\code{vector} of PSM-style modification annotations}
}
\value{
\code{vector} of Peptide-style modification annotations
}
\description{
Proteome Discoverer has a different style of annotation for
peptide modifications in PSM and peptide-level outputs. In PSM-level output,
each modification is listed separately, with the modification in parentheses
after the position. In peptide-level output, the modifications are summarised
more succintly, with the number of each modification followed by the positions
For example:
PSM: N-Term(Prot)(Acetyl); C2(Carbamidomethyl); C16(Carbamidomethyl)
Peptide:1xAcetyl \link{N-Term}; 2xCarbamidomethyl \link{C2; C16}
}
