% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbps.R
\docType{data}
\name{rbps}
\alias{rbps}
\title{RNA-binding proteins}
\format{
An object of class \code{character} of length 421.
}
\usage{
rbps
}
\description{
List of RNA binding proteins for use in LFQ vignette

\describe{
\item{column name}{\code{class}. Description.}
}
}
\keyword{datasets}
