% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_dots}
\alias{remove_dots}
\title{Remove duplicated full stops}
\usage{
remove_dots(x)
}
\arguments{
\item{x}{\code{character}. Contains duplicate full stops to be removed.}
}
\value{
Returns \code{character} with duplicate full stops removed.
}
\description{
A convenience function to remove any duplicated full stops
(aka periods) from the elements of a vector, e.g. will change \code{..} or \code{...}
or \code{....} etc. to just \code{.} Mainly used to fix column names.
}
\examples{

df <- data.frame(
  column...name = c(1, 2, 3)
)

colnames(df) <- remove_dots(colnames(df))

}
