% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{remove_silac_modifications}
\alias{remove_silac_modifications}
\title{Remove SILAC labels from modifications}
\usage{
remove_silac_modifications(mod_col, level = "psm")
}
\arguments{
\item{mod_col}{\verb{character vector} Modification column from Proteome Discoverer}

\item{level}{\code{character} Either 'psm' or 'peptide'}
}
\value{
\verb{character vector} updated Modifications column
}
\description{
This function removes SILAC heavy/light labels from a PD
Modifications column. When dealing with PSM level Proteome Discoverer output,
this is required to match PSMs for the same peptide, or to transfer PSM-level
features to peptide-level data.
}
