% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_x}
\alias{remove_x}
\title{Remove leading X}
\usage{
remove_x(x)
}
\arguments{
\item{x}{\code{character} or \code{string}.}
}
\value{
Returns \code{character} or \code{string} with leading X removed.
}
\description{
A convenience function to remove a leading capital X.
Is case sensitive.
}
\examples{

df <- data.frame('X1'=c(1,2))

removeX(colnames(df))

}
