% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{removeX}
\alias{removeX}
\title{Remove leading X}
\usage{
removeX(x)
}
\arguments{
\item{x}{\code{character}.}
}
\value{
Returns \code{character} with leading X removed
}
\description{
A convenience function to remove a leading X
}
\examples{

df <- data.frame('1'=c(1,2))

removeX(colnames(df))

}
