% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_features_per_protein.R
\name{restrict_features_per_protein}
\alias{restrict_features_per_protein}
\title{Remove features which are assigned to a protein with too few supporting
features in total}
\usage{
restrict_features_per_protein(
  obj,
  min_features,
  master_protein_col = "Master.Protein.Accessions",
  plot = TRUE
)
}
\arguments{
\item{obj}{\code{MSnSet} with PSM or peptide-level quantification}

\item{min_features}{\code{numeric} Threshold for minimum features per protein}

\item{master_protein_col}{\code{character} Column name for master protein}

\item{plot}{Set TRUE to plot histogram of features per protein per sample}
}
\value{
\code{MSnSet}
}
\description{
For summarisation of PSM or peptide to protein, we need a
minimum number of finite values per protein per sample. Where there are two
few finite values for a given protein in a given sample, this function will
replace all values for the protein features with NA. Proteins with two few
finite values in all samples will be removed entirely.

This function is useful for Label-Free Quantification but also when using
robust summarisation for isobaric tagging without PSM-level imputation
}
