% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{silac_psm_data}
\alias{silac_psm_data}
\title{PSM-level PD output from SILAC experiment}
\format{
A data frame with columns:
\describe{
\item{Quan.Channel}{Heavy of Light}
\item{Precursor.Abundance}{MS1 intensity}
\item{Annotated.Sequence}{Peptide sequence}
\item{Modifications}{Modifications}
\item{Isolation.Interference.in.Percent}{Interference/co-isolation}
}
}
\usage{
silac_psm_data
}
\description{
PSM-level PD output from SILAC experiment
}
\examples{
\dontrun{
 silac_psm_data
}
}
\keyword{datasets}
