% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{silac_psm_seq_int}
\alias{silac_psm_seq_int}
\title{Extract peptide sequencing and interference information from SILAC
PSM-level output}
\usage{
silac_psm_seq_int(obj, sequence_col = "Annotated.Sequence")
}
\arguments{
\item{obj}{\code{data.frame} PSM-level output from Proteome Discoverer}

\item{sequence_col}{\code{string} Column with peptide sequence}
}
\value{
\code{data.frame} indicating which SILAC peptides were MS2 sequenced and
the maximum interference across all PSMs for the peptide
}
\description{
Proteome Discoverer does not correctly propogate from PSM to
peptide level output which intensities are from "sequenced" peptides, e.g MS2
fragmentation. This information is useful to assess the accuracy of
quantification in peptides identified by mass shift relative to a sequenced
peptide in SILAC experiments
}
