% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fasta.R
\name{sub_crap}
\alias{sub_crap}
\title{Insert cRAP numbers into a character vector}
\usage{
sub_crap(x, start = 1, width = 3)
}
\arguments{
\item{x}{\verb{character vector}, each element must have two \code{|} symbols with
some text in between e.g. \verb{|sometext|}.}

\item{start}{\code{numeric}, the number to increment from, default is \code{1}}

\item{width}{\code{numeric}, how many digits the cRAP number should be, default is \code{3}}
}
\value{
Returns a \verb{character vector} the same length as x.
}
\description{
This function takes a character vector where each element is
roughly in the form of a UniProt header e.g. \code{"sp|XXXXXX|YYYY_YYYY Text goes here"}
and substitutes a cRAP number in place of the first \code{|} symbol, e.g.
\code{"sp|cRAP001|XXXXXX|YYYY_YYYY Text goes here"}.
}
\examples{
# basic use
sub_crap(c("|sometext|", "|moretext|"))

# start from a different number
sub_crap(c("|sometext|", "|moretext|"), start = 88)

# increase number width
sub_crap(c("|sometext|", "|moretext|"), start = 1111, width = 4)

}
