% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test function}
\usage{
test_function(data, x, y)
}
\arguments{
\item{data}{\code{data.frame}. Must have 2 numeric columns to plot.}

\item{x}{\code{numeric}. X values to plot.}

\item{y}{\code{numeric}. Y values to plot.}
}
\value{
ggplot using geom_point.
}
\description{
Generates a scatter plot with ggplot2.
}
\examples{
df <- data.frame(
  x = c(0, 1, 2),
  y = c(0, 1, 2)
)

test_function(df, x, y)
}
