% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_camprot}
\alias{theme_camprot}
\title{ggplot theme for package}
\usage{
theme_camprot(
  base_size = 20,
  base_family = "serif",
  aspect_square = TRUE,
  border = TRUE
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{aspect_square}{set aspect.ratio to 1}

\item{border}{include a border}
}
\value{
theme
}
\description{
A standard theme to use for plots generated by this package
}
