% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_csd.R
\name{theme_csd}
\alias{theme_csd}
\title{Charlotte's ggplot2 theme}
\usage{
theme_csd(
  base_size = 14,
  base_family = "sans",
  base_line_size = base_size/18,
  base_rect_size = base_size/9,
  aspect_square = TRUE,
  border = TRUE
)
}
\arguments{
\item{base_size}{\code{numeric}. Defines the overall size of the plot.}

\item{base_family}{\code{string}. Fontfamily to use, default is \code{"sans"}.}

\item{base_line_size}{\code{numeric}. Defines the line size to use.}

\item{base_rect_size}{\code{numeric}. Defines the rect size to use.}

\item{aspect_square}{\code{logical}. Should a square aspect ratio be used?
Default is \code{TRUE}.}

\item{border}{\code{logical}. Should the plot have a square border?
Default is \code{TRUE}.}
}
\value{
Returns a \emph{theme} object.
}
\description{
A standard theme to use for plots generated by this package.
}
\examples{
set.seed(11)
library(ggplot2)

df <- diamonds[sample(nrow(diamonds), 1000), ]

ggplot(df, aes(x = price, y = carat)) +
  geom_point() +
  theme_csd()

}
