% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmt_psm}
\alias{tmt_psm}
\title{PSM-level PD output from TMT experiment}
\format{
An MSnSet with fVarLabels:
\describe{
\item{Master.Protein.Accessions}{Protein Accessions}
\item{Isolation.Interference.in.Percent}{Interference/co-isolation}
\item{Average.Reporter.SN}{Average Signal:Noise for PMS}
}
}
\usage{
tmt_psm
}
\description{
PSM-level PD output from TMT experiment
}
\examples{
\dontrun{
 tmt_psm
}
}
\keyword{datasets}
