% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_TMT_PSMs.R
\name{update_average_sn}
\alias{update_average_sn}
\title{Adds new feature describing the average reporter Signal/Noise ratio.}
\usage{
update_average_sn(obj, sn_col = "Average.Reporter.SN")
}
\arguments{
\item{obj}{\code{MSnSet}. Contains PSMs for TMT quantification}

\item{sn_col}{\code{string}. Name of output column containing the average signal:noise.}
}
\value{
Returns an \code{MSnSet} with the average SN included as a new feature column.
}
\description{
PD column \code{Average.Reporter.SN} is NA when any tags have missing
values. This function adds an average reporter SN column which ignores missing
values. The assumption is that the intensity values in the \code{exprs} matrix
are Signal/Noise ratios, as reported by PD by default
}
