% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camprotR-package.R
\docType{package}
\name{camprotR-package}
\alias{camprotR}
\alias{camprotR-package}
\title{camprotR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Charlotte Dawson \email{csdaw@outlook.com} (\href{https://orcid.org/0000-0002-7151-5971}{ORCID})

Authors:
\itemize{
  \item Tom Smith \email{some@email.com} (\href{https://orcid.org/ORCID_goes_here}{ORCID})
}

}
\keyword{internal}
