% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{compare_sequences_per_protein}
\alias{compare_sequences_per_protein}
\title{Compare the number of peptides per protein pre/post-consensus master protein
assignment}
\usage{
compare_sequences_per_protein(seq2new_master_prot, seq_col, master_prot_col)
}
\arguments{
\item{seq2new_master_prot}{\code{data}. Output from merging
\code{\link{read_infiles}} & \code{\link{get_sequence2protein}}}

\item{seq_col}{\code{string} Column name for peptide sequence}

\item{master_prot_col}{\code{string} Column name for master protein(s)}
}
\value{
Returns an \code{ggplot} for the number of sequences per master protein
in the original infiles and after the consensus master protein assignment
}
\description{
Compare the number of peptides per protein pre/post-consensus master protein
assignment
}
\keyword{internal}
