% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{compare_single_master}
\alias{compare_single_master}
\title{Identify the proportion of peptides with a single assigned master protein}
\usage{
compare_single_master(seq2new_master_prot, master_prot_col)
}
\arguments{
\item{seq2new_master_prot}{\code{data}. Output from merging
\code{\link{read_infiles}} & \code{\link{get_sequence2protein}}}

\item{master_prot_col}{\code{string} Column name for master protein(s)}
}
\value{
Returns an \code{ggplot} for the number of sequences per master protein
in the original infiles and after the consensus master protein assignment
}
\description{
With object containing original and updated master protein
assignments, identify the proportion with a single assigned master protein
}
\keyword{internal}
