% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{estimate_incorporation}
\alias{estimate_incorporation}
\title{Estimate the incorporation rate directly from PD output}
\usage{
estimate_incorporation(
  psm_infile,
  peptide_infile,
  crap_fasta,
  master_protein_col = "Master.Protein.Accessions",
  protein_col = "Protein.Accessions",
  sequence_col = "Sequence",
  modifications_col = "Modifications",
  mix = 0,
  outdir
)
}
\arguments{
\item{psm_infile}{\code{string} filepath to psm-level PD output}

\item{peptide_infile}{\code{string} filepath to peptide-level PD output}

\item{crap_fasta}{\code{string}. filepath to cRAP fasta}

\item{master_protein_col}{\code{string}. Name of column containing master
proteins.}

\item{protein_col}{\code{string}. Name of column containing all protein
matches.}

\item{sequence_col}{\code{string}. Name of column containing peptide sequences}

\item{modifications_col}{\code{string}. Name of column containing peptide modifications}

\item{mix}{\code{numeric}. If Light material has been spiked in, what is the abundance
relative to the heavy material. E.g if they are equal, mix=1.
Default value = 0, e.g no Light spike in}

\item{outdir}{\code{string} filepath to directory for plots and incorporation summary table}
}
\value{
\code{incorporation}
}
\description{
SILAC incorporation can be estimated from peptide-level PD output
and summarised at peptide or protein level.
}
\details{
\strong{Peptide sequencing/mass shift identification}
In SILAC, Peptide identity of ions can be established through MS2 fragmentation
('peptide sequencing') or by mass shift (within tolerance limits) relative to
a sequenced peptide. Ideally, the correlation between heavy and light intensities
is the same regardless of whether light or heavy peptides are identified by
mass shift or sequencing. If e.g light peptide intensities are not well correlated
when they are identified by mass shift, this may indicate the mass shift identifications
are erroneously picking up 'ghost peptide', which will make incorporation estimation
difficult

\strong{Mixing Heavy and Light material for incorporation rate testing}
To get around the issue of 'ghost peptide', one can spike in Light material (
at cell or protein extract-level) to the Heavy material being analysed.
}
