% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{get_all_mappings}
\alias{get_all_mappings}
\title{Get all mappings for GO terms}
\usage{
get_all_mappings(go_ids, ontologies, verbose = TRUE, direction = "ancestor")
}
\arguments{
\item{go_ids}{\verb{character vector}. The GO.IDs to use}

\item{ontologies}{\verb{named character vector}. Names = GO.IDs and values =
Ontologies e.g. BP, CC, MF.}

\item{verbose}{\code{logical}.}

\item{direction}{\code{string} Either \code{"ancestor"} or
\code{"offspring"}.}
}
\value{
Returns a \verb{named list} of \verb{character vectors}. Names = GO.IDs and
values = all ancestor GO.IDs.
}
\description{
For a set of GO terms, obtain all of the ancestor or offspring terms.
}
\keyword{internal}
