% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cat_palette.R
\name{get_cat_palette}
\alias{get_cat_palette}
\title{Generate a colour-blind friendly palette for categorical colour encoding}
\usage{
get_cat_palette(n)
}
\arguments{
\item{n}{\code{numeric}. The number of colours required.}
}
\value{
Returns a \code{character} with the Hex codes for the colour palette.
}
\description{
For a given number of categories, identify a suitable
palette of colours which are colour-blind friendly. Palettes are derived from
\url{http://mkweb.bcgsc.ca/colorblind/palettes.mhtml#page-container}
}
