% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{get_incorporation}
\alias{get_incorporation}
\title{Determine the SILAC incorporation rate given intensity values for light and heavy}
\usage{
get_incorporation(light, heavy)
}
\arguments{
\item{light}{\code{numeric} Light peptide intensity}

\item{heavy}{\code{numeric} Heavy peptide intensity}
}
\value{
\code{incorporation}
}
\description{
SILAC incorporation is estimated by H / (H + L), where H=Heavy
and L=Light labelled peptide intensity. If L is NA, incorporation==1, if H is
NA, incorproration==0
}
