% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{get_parsimony_pep2prot}
\alias{get_parsimony_pep2prot}
\title{Identify the consensus peptide to master protein mapping across multiple runs}
\usage{
get_parsimony_pep2prot(
  infiles,
  seq_col = "Sequence",
  prot_col = "Protein.Accessions",
  master_prot_col = "Master.Protein.Accessions",
  compare_old_new = TRUE
)
}
\arguments{
\item{infiles}{\code{list}. Outfiles from PD}

\item{seq_col}{\code{string} Column name for peptide sequence}

\item{prot_col}{\code{string} Column name for protein(s)}

\item{master_prot_col}{\code{string} Column name for master protein(s)}

\item{compare_old_new}{\code{Logical} Compare the consensus master protein
assignments with the original assignments}
}
\value{
Returns an \code{data.frame} object which maps peptide sequences to
consensus master protein ID(s)
}
\description{
Wrapper function to read in infiles,
identify the parsimonious proteins to account for the peptides and
then (optionally) compare the new and old master protein assignments
}
