% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt_qc_plots.R
\name{get_psm_metrics}
\alias{get_psm_metrics}
\title{Summarise percentage of tag intensities below a given threshold and missing}
\usage{
get_psm_metrics(obj, threshold = 5.75, group_by_sample = FALSE)
}
\arguments{
\item{obj}{\code{MSnSet}.}

\item{threshold}{\code{numeric}. Minimum intensity threshold}

\item{group_by_sample}{\code{logical}. Group the summaries by sample}
}
\value{
\code{data.frame} object.
}
\description{
Given an \code{MSnSet}, return a \code{data.frame} summarising the
percentage of tag intensities below a given threshold, missing values
and the median tag intensity
}
