% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_runs.R
\name{get_sequence2protein}
\alias{get_sequence2protein}
\title{Approximate parsimony approach to map sequence to protein
across multiple samples}
\usage{
get_sequence2protein(seq2prot, seq_col, prot_col)
}
\arguments{
\item{seq2prot}{\code{data}. Output from \code{\link{readInfiles}}}

\item{seq_col}{\code{string} Column name for peptide sequence}

\item{prot_col}{\code{string} Column name for protein(s)}

\item{master_prot_col}{\code{string} Column name for master protein(s)}
}
\value{
Returns an \code{data.frame} object which uniquely maps peptide sequence
to master protein
}
\description{
Approximate parsimony approach to map sequence to protein
across multiple samples
}
\section{Uses an approximate parsimony approach (e.g least number of}{
proteins required to account for all observed sequences) to generate a
single map from sequence to protein for all samples
}

\keyword{internal}
