% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{plot_incorporation}
\alias{plot_incorporation}
\title{Plot annotated histogram of incorporation values}
\usage{
plot_incorporation(obj, level = "Peptide", mix = 0)
}
\arguments{
\item{obj}{\code{data.frame} containing features with column 'incorporation'.
If mix>0, must also contain column 'corrected_incorporation' where incorporation
is calculated using the corrected light peptide intensity}

\item{level}{\code{string} Name for feature level. e.g 'Peptide' or 'Protein'}
}
\value{
\code{list} with \code{p}=\code{ggplot} plot and \code{incorporation_estimates}=\code{list} of
incorporation estimates
}
\description{
Incorporation is estimated from the set of observed features. If
the test sample contains cells grown in 'heavy' media only, incorporation is
simply the mean observed incorporation. If the test sample contains a mixture of
'Heavy' and 'Light', the incorporation is best calculated from the median
incorporation using the corrected 'Light' intensity.
From a \code{data.frame}, with
}
