% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R, R/tmt_qc_plots.R
\name{plot_TMT_notch}
\alias{plot_TMT_notch}
\title{Plot histograms for TMT tag intensities per sample}
\usage{
plot_TMT_notch(
  obj,
  notch_lower = 3.75,
  notch_upper = 5.75,
  facet_by_sample = FALSE
)

plot_TMT_notch(
  obj,
  notch_lower = 3.75,
  notch_upper = 5.75,
  facet_by_sample = FALSE
)
}
\arguments{
\item{obj}{\code{MSnSet} containing PSM level TMT intensities}

\item{notch_lower}{\code{numeric}. Lower boundary of notch.}

\item{notch_upper}{\code{numeric}. Upper boundary of notch.}

\item{facet_by_sample}{\code{logical}. Facet the plot by sample.}
}
\value{
\code{ggplot} object.

\code{ggplot} object.
}
\description{
Given an \code{MSnSet}, return a histogram
of TMT tag intensities per sample, annotated with the proportion of
intensities below the upper boundary of the notch region

The distribution of TMT intensities obtained by Orbitrap has
been observed to have a 'notch' where few values are observed. This function
returns a histogram of TMT tag intensities per sample, annotated with the
proportion of intensities below the upper boundary of the notch region
}
