% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfq.R
\name{read_infiles}
\alias{read_infiles}
\title{Extract sequence:protein ID mappings}
\usage{
read_infiles(infiles, seq_col, prot_col, master_prot_col)
}
\arguments{
\item{infiles}{\code{list}. Outfiles from PD}

\item{seq_col}{\code{string} Column name for peptide sequence}

\item{prot_col}{\code{string} Column name for protein(s)}

\item{master_prot_col}{\code{string} Column name for master protein(s)}
}
\value{
Returns an \code{data.frame} object which maps peptide sequences to
protein and master protein ID(s) for each infile
}
\description{
Given a list of infiles, read the files,
retain the sequence and protein ID columns and bind the rows
}
\keyword{internal}
