% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silac.R
\name{remove_silac_modifications}
\alias{remove_silac_modifications}
\title{Remove SILAC labels from modifications}
\usage{
remove_silac_modifications(obj, level = "psm")
}
\arguments{
\item{obj}{\code{data.frame} PSM or Peptide-level output from Proteome Discoverer}

\item{level}{\code{string} Eiher 'psm' or 'peptide'}
}
\value{
\code{data.frame} with the updated Modifications column
}
\description{
This function removes SILAC heavy/light labels from the
Modifications column. When dealing with PSM level Proteome Discoverer output,
this is required to match PSMs for the same peptide, or to transfer PSM-level
features to peptide-level data.
}
